requirements.txt
python-telegram-bot==20.8
openai==1.58.1
python-docx==1.1.2
python-dotenv==1.0.1
sqlalchemy==2.0.36
yookassa==3.5.0
psutil==5.9.8
